graphql queries extracted do model files
