+++ src/app/components/comments/comments-page.model.ts
@@ -1,0 +1,65 @@
+import gql from 'graphql-tag';
+
+export const commentQuery = gql`
+  query Comment($repoName: String!) {
+    # Eventually move this into a no fetch query right on the entry
+    # since we literally just need this info to determine whether to
+    # show upvote/downvote buttons
+    currentUser {
+      login
+      html_url
+    }
+    entry(repoFullName: $repoName) {
+      id
+      postedBy {
+        login
+        html_url
+      }
+      createdAt
+      comments {
+        id
+        postedBy {
+          login
+          html_url
+        }
+        createdAt
+        content
+      }
+      repository {
+        full_name
+        html_url
+        description
+        open_issues_count
+        stargazers_count
+      }
+    }
+  }
+`;
+
+export const subscriptionQuery = gql`
+  subscription onCommentAdded($repoFullName: String!){
+    commentAdded(repoFullName: $repoFullName){
+      id
+      postedBy {
+        login
+        html_url
+      }
+      createdAt
+      content
+    }
+  }
+`;
+
+export const submitCommentMutation = gql`
+  mutation submitComment($repoFullName: String!, $commentContent: String!) {
+    submitComment(repoFullName: $repoFullName, commentContent: $commentContent) {
+      id
+      postedBy {
+        login
+        html_url
+      }
+      createdAt
+      content
+    }
+  }
+`;
