+++ src/app/components/feed/feed.component.html
@@ -1,11 +1,25 @@
 <loading *ngIf="loading"></loading>
 <div *ngIf="!loading">
-  <feed-entry
-    *ngFor="let entry of feed"
-    [entry]="entry"
-    [currentUser]="currentUser"
-    (onVote)="onVote($event)">
-  </feed-entry>
 
-  <a (click)="fetchMore()">Load more</a>
+  <div class="form-group">
+    <input
+      type="text"
+      class="form-control"
+      id="searchInput"
+      placeholder="Search..."
+      name="searchInput"
+      required
+    />
+  </div>
+
+  <div>
+    <feed-entry
+      *ngFor="let entry of feed"
+      [entry]="entry"
+      [currentUser]="currentUser"
+      (onVote)="onVote($event)">
+    </feed-entry>
+
+    <a (click)="fetchMore()">Load more</a>
+  </div>
 </div>
