+++ src/app/components/new-entry.model.ts
@@ -1,0 +1,9 @@
+import gql from 'graphql-tag';
+
+export const submitRepositoryMutation = gql`
+  mutation submitRepository($repoFullName: String!) {
+    submitRepository(repoFullName: $repoFullName) {
+      createdAt
+    }
+  }
+`;
