+++ src/app/components/profile.component.ts
@@ -1,17 +1,7 @@
 import {Component, OnInit, OnDestroy} from '@angular/core';
 import {Angular2Apollo} from 'angular2-apollo';
 import {Subscription} from 'rxjs/Subscription';
-
-import gql from 'graphql-tag';
-
-const CurrentUserQuery = gql`
-  query CurrentUserForProfile {
-    currentUser {
-      login
-      avatar_url
-    }
-  }
-`;
+import {CurrentUserQuery} from './profile.model';
 
 @Component({
   selector: 'profile',
