+++ src/app/components/profile.model.ts
@@ -1,0 +1,10 @@
+import gql from 'graphql-tag';
+
+export const CurrentUserQuery = gql`
+  query CurrentUserForProfile {
+    currentUser {
+      login
+      avatar_url
+    }
+  }
+`;
